Adding authorization: adding ref to the Model, adding User Id to Post…

+++ backend/routes/posts.js
@@ -36,8 +36,11 @@ router.post(
     const post = new Post({
         title: req.body.title,
         content: req.body.content,
-        imagePath: url + "/images/" + req.file.filename
+        imagePath: url + "/images/" + req.file.filename,
+        creator: req.userData.userId
     });
+    
+    res.status(200).json({});
     post.save().then(createdPost => {
         res.status(201).json({
             message: 'Post added successfully !',
@@ -101,23 +104,28 @@ router.put(
         _id: req.body.id,
         title: req.body.title,
         content: req.body.content,
-        imagePath: imagePath
+        imagePath: imagePath,
+        creator: req.userData.userId
     });
-    console.log('post', post)
-    Post.updateOne({_id: req.params.id}, post)
-    .then(result => {
-        console.log('result', result);
-        res.status(200).json({
-            message: 'Post updated sucessfully !'
-        });
+  
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then(result => {
+        if (result.modifiedCount > 0) {
+            res.status(200).json({ message: 'Post updated sucessfully !' });
+        } else {
+            res.status(401).json({ message: 'Not authorized to update post !' });
+        };
     });
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-    Post.deleteOne({ _id: req.params.id}).then(result => {
+    Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then(result => {
         console.log('result', result)
-        res.status(200).json({ message: 'Post deleted !'});
-    })
+        if (result.deletedCount > 0) {
+            res.status(200).json({ message: 'Post deletion sucessfully !'});
+        } else {
+            res.status(401).json({ message: 'Not authorized to delete post !' });
+        };
+    });
 });
 
 module.exports = router;
